Registration workflow confirmed
